convert contextualize to TS
