+++ packages/node/src/bugsnag.ts
@@ -17,7 +17,7 @@ import pluginNodeDevice from '@bugsnag/plugin-node-device'
 import pluginNodeUncaughtException from '@bugsnag/plugin-node-uncaught-exception'
 import pluginNodeUnhandledRejection from '@bugsnag/plugin-node-unhandled-rejection'
 import pluginIntercept from '@bugsnag/plugin-intercept'
-import pluginContextualize from '@bugsnag/plugin-contextualize'
+import pluginContextualize from'@bugsnag/plugin-contextualize'
 import pluginStackframePathNormaliser from '@bugsnag/plugin-stackframe-path-normaliser'
 import pluginConsoleBreadcrumbs from '@bugsnag/plugin-console-breadcrumbs'
 import { BugsnagStatic, Client, Config, Event, Logger } from '@bugsnag/core'
