+++ packages/plugin-contextualize/rollup.config.npm.mjs
@@ -1,0 +1,6 @@
+import createRollupConfig from "../../.rollup/index.mjs"
+
+export default createRollupConfig({
+    input: "src/contextualize.ts",
+    external: [/node_modules/],
+})
