+++ packages/plugin-contextualize/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "include": ["src/**/*.ts"],
+  "compilerOptions": {
+    "lib": ["es2017"],
+    "types": ["node"]
+  }
+}
